Extract war-room storage into a shared package
`web/src/server/war-room.ts` and `redis-lock.ts` are pure-Redis
libraries (only `@upstash/redis` and `crypto` imports) — they live
under `web/` only by historical accident. Bot will need the same
storage layer for the multi-tenant queen migration (drop
HIVEMOOT_BOT_AGENT_TOKEN, write Redis directly), so move them to a
shared location both packages can import.

Layout:
- New `shared/war-room/` package, source-only, no own node_modules.
- Web installs it via `file:../shared/war-room` — npm creates a
  symlink at `web/node_modules/@hivemoot/war-room` pointing at the
  source. Single install of `@upstash/redis` (web's), no
  dual-package hazard.
- `tsconfig.json`: `preserveSymlinks: true` so type identity stays
  consistent across the symlink boundary.
- `next.config.ts`: `transpilePackages: ["@hivemoot/war-room"]` so
  Turbopack runs the shared `.ts` files through the same transform
  pipeline as web's source. `outputFileTracingRoot` extends Next's
  bundle scope to the monorepo root.
- `vitest.config.ts`: `resolve.alias` mirrors the package name and
  the `include` glob picks up shared's tests so they run in web's
  vitest pass.

Pure refactor — no behavior changes. All 1380 tests pass; bot's
1954 tests unaffected.
