+++ shared/war-room/src/index.ts
@@ -1,0 +1,7 @@
+// Re-export the war-room storage layer and redis-lock helper from a
+// single entrypoint. Consumers can also reach `redis-lock` via the
+// `./redis-lock` subpath export (see package.json) when they only
+// need the lock primitive.
+
+export * from "./war-room";
+export * from "./redis-lock";
