+++ shared/war-room/src/war-room.ts
