+++ web/src/app/api/rooms/[roomId]/contributions/route.test.ts
@@ -10,9 +10,9 @@ vi.mock("@/server/agent-token-v1-auth", () => ({
   authenticateAgentRequestV1: vi.fn(),
 }));
 
-vi.mock("@/server/war-room", async () => {
-  const real = await vi.importActual<typeof import("@/server/war-room")>(
-    "@/server/war-room",
+vi.mock("@hivemoot/war-room", async () => {
+  const real = await vi.importActual<typeof import("@hivemoot/war-room")>(
+    "@hivemoot/war-room",
   );
   return {
     ...real,
@@ -36,7 +36,7 @@ import {
   RoomParticipantStatePreconditionError,
   RoomContributionTooLargeError,
   ContributionValidationError,
-} from "@/server/war-room";
+} from "@hivemoot/war-room";
 import { GET, POST, DELETE } from "./route";
 
 const mockedAuth = vi.mocked(authenticateAgentRequestV1);
