+++ web/src/app/api/rooms/[roomId]/participants/route.ts
@@ -18,7 +18,7 @@ import {
   getRoomParticipants,
   RoomNotFoundError,
   RoomIdFormatError,
-} from "@/server/war-room";
+} from "@hivemoot/war-room";
 
 export async function GET(
   request: NextRequest,
