+++ web/src/app/api/rooms/[roomId]/present/route.test.ts
@@ -9,9 +9,9 @@ vi.mock("@/server/agent-token-v1-auth", () => ({
   authenticateAgentRequestV1: vi.fn(),
 }));
 
-vi.mock("@/server/war-room", async () => {
-  const real = await vi.importActual<typeof import("@/server/war-room")>(
-    "@/server/war-room",
+vi.mock("@hivemoot/war-room", async () => {
+  const real = await vi.importActual<typeof import("@hivemoot/war-room")>(
+    "@hivemoot/war-room",
   );
   return {
     ...real,
@@ -27,7 +27,7 @@ import {
   RoomEventStatusPreconditionError,
   RoomEventBodyTooLargeError,
   RoomParticipantOwnerConflictError,
-} from "@/server/war-room";
+} from "@hivemoot/war-room";
 import { POST } from "./route";
 
 const mockedAuth = vi.mocked(authenticateAgentRequestV1);
