+++ web/tsconfig.json
@@ -15,6 +15,7 @@
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
+    "preserveSymlinks": true,
     "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
